+++ CHANGES.md
@@ -9,6 +9,7 @@
 ##### Additions :tada:
 
 - Added `Cesium3DTileStyle.fromUrl` for loading a style from a url. [#10348](https://github.com/CesiumGS/cesium/pull/10348)
+- Added `IndexDatatype.fromTypedArray`. [#10350](https://github.com/CesiumGS/cesium/pull/10350)
 
 ##### Fixes :wrench: